+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 2026-04-08
+- #2695 adds a new `ext_getStorageProof` RPC endpoint to the EVM module
 - #2704 Replaces `reth`-related crates. No breakage is intended.
 
 # 2026-04-02